Load ib_ipoib kernel module on boot in OFED & OpenHPC images
Use the systemd-modules-load element from stackhpc-image-elements to load the
ib_ipoib kernel module on boot in the CentOS7.4-OFED4.1 and CentOS7.4-OpenHPC
images. This ensures that the Infiniband IP over IB interface is present when
cloud-init processes the config-drive. Using a config-drive allows us to
allocate IP addresses on the IB network without connectivity on the neutron
network node.